Unifying non-relational implementations
