+++ lisa/lisa-analyses/imp-testcases/non-interference/confidentiality/report.json
@@ -23,14 +23,14 @@
   "files" : [ "report.json", "untyped_Confidentiality.confidentiality_lec16(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec17(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec18ex1(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec18ex2(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec19ex1(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec19ex2(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec19ex3(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec19ex4(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec19ex5(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec5ex1(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec5ex2(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec6(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec7(Confidentiality__this).json", "untyped_Confidentiality.confidentiality_lec8(Confidentiality__this).json" ],
   "info" : {
     "cfgs" : "14",
-    "duration" : "88ms",
-    "end" : "2023-09-07T11:45:54.614+02:00",
+    "duration" : "486ms",
+    "end" : "2025-01-31T12:39:50.549+01:00",
     "expressions" : "130",
     "files" : "14",
     "globals" : "0",
     "members" : "14",
     "programs" : "1",
-    "start" : "2023-09-07T11:45:54.526+02:00",
+    "start" : "2025-01-31T12:39:50.063+01:00",
     "statements" : "70",
     "units" : "1",
     "version" : "0.1b8",
