+++ lisa/lisa-analyses/imp-testcases/non-interference/confidentiality/untyped_Confidentiality.confidentiality_lec17(Confidentiality__this).json
@@ -1,1 +1,1 @@
-{"name":"untyped Confidentiality::confidentiality_lec17(Confidentiality* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":"l = +(h, l)"},{"id":7,"text":"l"},{"id":8,"subNodes":[9,10],"text":"+(h, l)"},{"id":9,"text":"h"},{"id":10,"text":"l"},{"id":11,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":11,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["h + l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":10,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":11,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}}]}
+{"name":"untyped Confidentiality::confidentiality_lec17(Confidentiality* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":"l = +(h, l)"},{"id":7,"text":"l"},{"id":8,"subNodes":[9,10],"text":"+(h, l)"},{"id":9,"text":"h"},{"id":10,"text":"l"},{"id":11,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":11,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":"#TOP#"}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":"#TOP#"}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["h + l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":10,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":11,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}}]}
