+++ lisa/lisa-analyses/imp-testcases/non-interference/confidentiality/untyped_Confidentiality.confidentiality_lec18ex1(Confidentiality__this).json
@@ -1,1 +1,1 @@
-{"name":"untyped Confidentiality::confidentiality_lec18ex1(Confidentiality* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"<=(h, 30)"},{"id":4,"text":"h"},{"id":5,"text":"30"},{"id":6,"subNodes":[7,8],"text":"h = 5"},{"id":7,"text":"h"},{"id":8,"text":"5"},{"id":9,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"TrueEdge"},{"sourceId":3,"destId":9,"kind":"FalseEdge"},{"sourceId":6,"destId":9,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":3,"description":{"expressions":["h <= 30"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["30"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"HL"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"HL"}}}},{"nodeId":8,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"HL"}}}},{"nodeId":9,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}}]}
+{"name":"untyped Confidentiality::confidentiality_lec18ex1(Confidentiality* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"<=(h, 30)"},{"id":4,"text":"h"},{"id":5,"text":"30"},{"id":6,"subNodes":[7,8],"text":"h = 5"},{"id":7,"text":"h"},{"id":8,"text":"5"},{"id":9,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"TrueEdge"},{"sourceId":3,"destId":9,"kind":"FalseEdge"},{"sourceId":6,"destId":9,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":"#TOP#"}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":"#TOP#"}}},{"nodeId":3,"description":{"expressions":["h <= 30"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["30"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"HL"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"HL"}}}},{"nodeId":8,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"HL"}}}},{"nodeId":9,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}}]}
