+++ lisa/lisa-analyses/imp-testcases/non-interference/confidentiality/untyped_Confidentiality.confidentiality_lec18ex2(Confidentiality__this).json
@@ -1,1 +1,1 @@
-{"name":"untyped Confidentiality::confidentiality_lec18ex2(Confidentiality* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":"<=(h, 30)"},{"id":7,"text":"h"},{"id":8,"text":"30"},{"id":9,"subNodes":[10,11],"text":"l = 5"},{"id":10,"text":"l"},{"id":11,"text":"5"},{"id":12,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":9,"kind":"TrueEdge"},{"sourceId":6,"destId":12,"kind":"FalseEdge"},{"sourceId":9,"destId":12,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h <= 30"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["30"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"HL"}}}},{"nodeId":10,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"HL"}}}},{"nodeId":11,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"HL"}}}},{"nodeId":12,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}}]}
+{"name":"untyped Confidentiality::confidentiality_lec18ex2(Confidentiality* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":"<=(h, 30)"},{"id":7,"text":"h"},{"id":8,"text":"30"},{"id":9,"subNodes":[10,11],"text":"l = 5"},{"id":10,"text":"l"},{"id":11,"text":"5"},{"id":12,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":9,"kind":"TrueEdge"},{"sourceId":6,"destId":12,"kind":"FalseEdge"},{"sourceId":9,"destId":12,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":"#TOP#"}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Confidentiality*"]},"value":"#TOP#"}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h <= 30"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["30"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"HL"}}}},{"nodeId":10,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"HL"}}}},{"nodeId":11,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"HL"}}}},{"nodeId":12,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Confidentiality*"]},"value":{"map":{"h":"HL","l":"LL"},"state":"LH"}}}}]}
