+++ lisa/lisa-analyses/imp-testcases/non-interference/integrity/report.json
@@ -23,14 +23,14 @@
   "files" : [ "report.json", "untyped_Integrity.integrity_lec16(Integrity__this).json", "untyped_Integrity.integrity_lec17(Integrity__this).json", "untyped_Integrity.integrity_lec18ex1(Integrity__this).json", "untyped_Integrity.integrity_lec18ex2(Integrity__this).json", "untyped_Integrity.integrity_lec19ex1(Integrity__this).json", "untyped_Integrity.integrity_lec19ex2(Integrity__this).json", "untyped_Integrity.integrity_lec19ex3(Integrity__this).json", "untyped_Integrity.integrity_lec19ex4(Integrity__this).json", "untyped_Integrity.integrity_lec19ex5(Integrity__this).json", "untyped_Integrity.integrity_lec5ex1(Integrity__this).json", "untyped_Integrity.integrity_lec5ex2(Integrity__this).json", "untyped_Integrity.integrity_lec6(Integrity__this).json", "untyped_Integrity.integrity_lec7(Integrity__this).json", "untyped_Integrity.integrity_lec8(Integrity__this).json" ],
   "info" : {
     "cfgs" : "14",
-    "duration" : "168ms",
-    "end" : "2023-09-07T11:45:53.865+02:00",
+    "duration" : "555ms",
+    "end" : "2025-01-31T12:39:02.180+01:00",
     "expressions" : "130",
     "files" : "14",
     "globals" : "0",
     "members" : "14",
     "programs" : "1",
-    "start" : "2023-09-07T11:45:53.697+02:00",
+    "start" : "2025-01-31T12:39:01.625+01:00",
     "statements" : "70",
     "units" : "1",
     "version" : "0.1b8",
