+++ lisa/lisa-analyses/imp-testcases/non-interference/integrity/untyped_Integrity.integrity_lec16(Integrity__this).json
@@ -1,1 +1,1 @@
-{"name":"untyped Integrity::integrity_lec16(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":"h = +(l, 4)"},{"id":7,"text":"h"},{"id":8,"subNodes":[9,10],"text":"+(l, 4)"},{"id":9,"text":"l"},{"id":10,"text":"4"},{"id":11,"subNodes":[12,13],"text":"l = -(l, 3)"},{"id":12,"text":"l"},{"id":13,"subNodes":[14,15],"text":"-(l, 3)"},{"id":14,"text":"l"},{"id":15,"text":"3"},{"id":16,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":11,"kind":"SequentialEdge"},{"sourceId":11,"destId":16,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["l + 4"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":10,"description":{"expressions":["4"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":11,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":12,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":13,"description":{"expressions":["l - 3"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":14,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":15,"description":{"expressions":["3"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":16,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}}]}
+{"name":"untyped Integrity::integrity_lec16(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":"h = +(l, 4)"},{"id":7,"text":"h"},{"id":8,"subNodes":[9,10],"text":"+(l, 4)"},{"id":9,"text":"l"},{"id":10,"text":"4"},{"id":11,"subNodes":[12,13],"text":"l = -(l, 3)"},{"id":12,"text":"l"},{"id":13,"subNodes":[14,15],"text":"-(l, 3)"},{"id":14,"text":"l"},{"id":15,"text":"3"},{"id":16,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":11,"kind":"SequentialEdge"},{"sourceId":11,"destId":16,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["l + 4"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":10,"description":{"expressions":["4"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":11,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":12,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":13,"description":{"expressions":["l - 3"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":14,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":15,"description":{"expressions":["3"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":16,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}}]}
