+++ lisa/lisa-analyses/imp-testcases/non-interference/integrity/untyped_Integrity.integrity_lec5ex1(Integrity__this).json
@@ -1,1 +1,1 @@
-{"name":"untyped Integrity::integrity_lec5ex1(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = h"},{"id":4,"text":"l"},{"id":5,"text":"h"},{"id":6,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}}]}
+{"name":"untyped Integrity::integrity_lec5ex1(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = h"},{"id":4,"text":"l"},{"id":5,"text":"h"},{"id":6,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}}]}
