+++ lisa/lisa-analyses/imp-testcases/non-interference/integrity/untyped_Integrity.integrity_lec6(Integrity__this).json
@@ -1,1 +1,1 @@
-{"name":"untyped Integrity::integrity_lec6(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":">(h, 5000)"},{"id":7,"text":"h"},{"id":8,"text":"5000"},{"id":9,"subNodes":[10,11],"text":"l = 0"},{"id":10,"text":"l"},{"id":11,"text":"0"},{"id":12,"subNodes":[13,14],"text":"l = 1"},{"id":13,"text":"l"},{"id":14,"text":"1"},{"id":15,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":9,"kind":"TrueEdge"},{"sourceId":6,"destId":12,"kind":"FalseEdge"},{"sourceId":9,"destId":15,"kind":"SequentialEdge"},{"sourceId":12,"destId":15,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h > 5000"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["5000"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":10,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":11,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":12,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":13,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":14,"description":{"expressions":["1"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":15,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}}]}
+{"name":"untyped Integrity::integrity_lec6(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1,2],"text":"h = 5"},{"id":1,"text":"h"},{"id":2,"text":"5"},{"id":3,"subNodes":[4,5],"text":"l = 0"},{"id":4,"text":"l"},{"id":5,"text":"0"},{"id":6,"subNodes":[7,8],"text":">(h, 5000)"},{"id":7,"text":"h"},{"id":8,"text":"5000"},{"id":9,"subNodes":[10,11],"text":"l = 0"},{"id":10,"text":"l"},{"id":11,"text":"0"},{"id":12,"subNodes":[13,14],"text":"l = 1"},{"id":13,"text":"l"},{"id":14,"text":"1"},{"id":15,"text":"ret"}],"edges":[{"sourceId":0,"destId":3,"kind":"SequentialEdge"},{"sourceId":3,"destId":6,"kind":"SequentialEdge"},{"sourceId":6,"destId":9,"kind":"TrueEdge"},{"sourceId":6,"destId":12,"kind":"FalseEdge"},{"sourceId":9,"destId":15,"kind":"SequentialEdge"},{"sourceId":12,"destId":15,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":1,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":2,"description":{"expressions":["5"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":3,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":4,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":5,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL"},"state":"LH"}}}},{"nodeId":6,"description":{"expressions":["h > 5000"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":7,"description":{"expressions":["h"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":8,"description":{"expressions":["5000"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}},{"nodeId":9,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":10,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":11,"description":{"expressions":["0"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":12,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":13,"description":{"expressions":["l"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":14,"description":{"expressions":["1"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"HL"}}}},{"nodeId":15,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"h":["int32"],"l":["int32"],"this":["Integrity*"]},"value":{"map":{"h":"HL","l":"HH"},"state":"LH"}}}}]}
