+++ lisa/lisa-analyses/imp-testcases/non-interference/interproc/report.json
@@ -9,14 +9,14 @@
   "files" : [ "report.json", "untyped_Integrity.confidentiality_lec5ex1_declassified(Integrity__this)_-791177979.json", "untyped_Integrity.confidentiality_lec6_declassified(Integrity__this)_-791180831.json", "untyped_Integrity.declassify(Integrity__this,_untyped_y)_451999607.json", "untyped_Integrity.declassify(Integrity__this,_untyped_y)_452087709.json", "untyped_Integrity.main(Integrity__this).json" ],
   "info" : {
     "cfgs" : "4",
-    "duration" : "31ms",
-    "end" : "2023-09-07T11:45:55.211+02:00",
+    "duration" : "571ms",
+    "end" : "2025-01-31T12:41:39.010+01:00",
     "expressions" : "29",
     "files" : "5",
     "globals" : "0",
     "members" : "4",
     "programs" : "1",
-    "start" : "2023-09-07T11:45:55.180+02:00",
+    "start" : "2025-01-31T12:41:38.439+01:00",
     "statements" : "17",
     "units" : "1",
     "version" : "0.1b8",
