+++ lisa/lisa-analyses/imp-testcases/non-interference/interproc/untyped_Integrity.main(Integrity__this).json
@@ -1,1 +1,1 @@
-{"name":"untyped Integrity::main(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1],"text":"confidentiality_lec5ex1_declassified(this)"},{"id":1,"text":"this"},{"id":2,"subNodes":[3],"text":"confidentiality_lec6_declassified(this)"},{"id":3,"text":"this"},{"id":4,"text":"ret"}],"edges":[{"sourceId":0,"destId":2,"kind":"SequentialEdge"},{"sourceId":2,"destId":4,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":1,"description":{"expressions":["this"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":2,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":3,"description":{"expressions":["this"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}},{"nodeId":4,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":{"map":"empty","state":"LH"}}}}]}
+{"name":"untyped Integrity::main(Integrity* this)","description":null,"nodes":[{"id":0,"subNodes":[1],"text":"confidentiality_lec5ex1_declassified(this)"},{"id":1,"text":"this"},{"id":2,"subNodes":[3],"text":"confidentiality_lec6_declassified(this)"},{"id":3,"text":"this"},{"id":4,"text":"ret"}],"edges":[{"sourceId":0,"destId":2,"kind":"SequentialEdge"},{"sourceId":2,"destId":4,"kind":"SequentialEdge"}],"descriptions":[{"nodeId":0,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":1,"description":{"expressions":["this"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":2,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":3,"description":{"expressions":["this"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}},{"nodeId":4,"description":{"expressions":["skip"],"state":{"heap":"monolith","type":{"this":["Integrity*"]},"value":"#TOP#"}}}]}
