+++ lisa/lisa-analyses/src/main/java/it/unive/lisa/LiSAFactory.java
@@ -1,22 +1,5 @@
 package it.unive.lisa;
 
-import it.unive.lisa.analysis.AbstractState;
-import it.unive.lisa.analysis.dataflow.DataflowElement;
-import it.unive.lisa.analysis.dataflow.DefiniteDataflowDomain;
-import it.unive.lisa.analysis.dataflow.PossibleDataflowDomain;
-import it.unive.lisa.analysis.heap.HeapDomain;
-import it.unive.lisa.analysis.nonrelational.heap.HeapEnvironment;
-import it.unive.lisa.analysis.nonrelational.heap.NonRelationalHeapDomain;
-import it.unive.lisa.analysis.nonrelational.inference.InferenceSystem;
-import it.unive.lisa.analysis.nonrelational.inference.InferredValue;
-import it.unive.lisa.analysis.nonrelational.value.NonRelationalTypeDomain;
-import it.unive.lisa.analysis.nonrelational.value.NonRelationalValueDomain;
-import it.unive.lisa.analysis.nonrelational.value.TypeEnvironment;
-import it.unive.lisa.analysis.nonrelational.value.ValueEnvironment;
-import it.unive.lisa.analysis.type.TypeDomain;
-import it.unive.lisa.analysis.value.ValueDomain;
-import it.unive.lisa.interprocedural.InterproceduralAnalysis;
-import it.unive.lisa.interprocedural.callgraph.CallGraph;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
@@ -31,10 +14,27 @@ import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.reflections.Reflections;
 import org.reflections.scanners.SubTypesScanner;
+
+import it.unive.lisa.analysis.AbstractState;
+import it.unive.lisa.analysis.dataflow.DataflowElement;
+import it.unive.lisa.analysis.dataflow.DefiniteDataflowDomain;
+import it.unive.lisa.analysis.dataflow.PossibleDataflowDomain;
+import it.unive.lisa.analysis.heap.HeapDomain;
+import it.unive.lisa.analysis.nonrelational.heap.HeapEnvironment;
+import it.unive.lisa.analysis.nonrelational.heap.NonRelationalHeapDomain;
+import it.unive.lisa.analysis.nonrelational.value.NonRelationalTypeDomain;
+import it.unive.lisa.analysis.nonrelational.value.NonRelationalValueDomain;
+import it.unive.lisa.analysis.nonrelational.value.TypeEnvironment;
+import it.unive.lisa.analysis.nonrelational.value.ValueEnvironment;
+import it.unive.lisa.analysis.type.TypeDomain;
+import it.unive.lisa.analysis.value.ValueDomain;
+import it.unive.lisa.interprocedural.InterproceduralAnalysis;
+import it.unive.lisa.interprocedural.callgraph.CallGraph;
 
 /**
  * An utility class for instantiating analysis components, that is, modular
@@ -126,8 +126,6 @@ public final class LiSAFactory {
 			return true;
 		else if (NonRelationalTypeDomain.class.isAssignableFrom(actual) && desired.isAssignableFrom(TypeDomain.class))
 			return true;
-		else if (InferredValue.class.isAssignableFrom(actual) && desired.isAssignableFrom(ValueDomain.class))
-			return true;
 		else if (DataflowElement.class.isAssignableFrom(actual) && desired.isAssignableFrom(ValueDomain.class))
 			return true;
 		else
@@ -143,8 +141,6 @@ public final class LiSAFactory {
 			return new ValueEnvironment((NonRelationalValueDomain<?>) param);
 		else if (NonRelationalTypeDomain.class.isAssignableFrom(param.getClass()))
 			return new TypeEnvironment((NonRelationalTypeDomain<?>) param);
-		else if (InferredValue.class.isAssignableFrom(param.getClass()))
-			return new InferenceSystem((InferredValue<?>) param);
 		else if (DataflowElement.class.isAssignableFrom(param.getClass())) {
 			Class<? extends DataflowElement> elem = (Class<? extends DataflowElement>) param.getClass();
 			if (elem.getGenericInterfaces().length == 0)
@@ -313,7 +309,6 @@ public final class LiSAFactory {
 		in.add(new ConfigurableComponent(NonRelationalHeapDomain.class));
 		in.add(new ConfigurableComponent(NonRelationalValueDomain.class));
 		in.add(new ConfigurableComponent(NonRelationalTypeDomain.class));
-		in.add(new ConfigurableComponent(InferredValue.class));
 		in.add(new ConfigurableComponent(DataflowElement.class));
 		return in;
 	}
