+++ lisa/lisa-analyses/src/main/java/it/unive/lisa/analysis/numeric/Interval.java
@@ -453,7 +453,8 @@ public class Interval implements BaseNonRelationalValueDomain<Interval>, Compara
 
 		Interval update = null;
 		if (operator == ComparisonEq.INSTANCE)
-			update = eval;
+			// if eval is not a possible value, we go to bottom
+			update = starting.glb(eval); 
 		else if (operator == ComparisonGe.INSTANCE)
 			if (rightIsExpr)
 				update = lowIsMinusInfinity ? null : starting.glb(low_inf);
