+++ lisa/lisa-analyses/src/test/java/it/unive/lisa/cron/InformationFlowTest.java
@@ -1,5 +1,9 @@
 package it.unive.lisa.cron;
 
+import java.util.Collection;
+
+import org.junit.Test;
+
 import it.unive.lisa.AnalysisTestExecutor;
 import it.unive.lisa.CronConfiguration;
 import it.unive.lisa.DefaultConfiguration;
@@ -9,7 +13,7 @@ import it.unive.lisa.analysis.SemanticException;
 import it.unive.lisa.analysis.SimpleAbstractState;
 import it.unive.lisa.analysis.heap.MonolithicHeap;
 import it.unive.lisa.analysis.nonInterference.NonInterference;
-import it.unive.lisa.analysis.nonrelational.inference.InferenceSystem;
+import it.unive.lisa.analysis.nonInterference.NonInterference.NI;
 import it.unive.lisa.analysis.nonrelational.value.TypeEnvironment;
 import it.unive.lisa.analysis.nonrelational.value.ValueEnvironment;
 import it.unive.lisa.analysis.taint.BaseTaint;
@@ -33,8 +37,6 @@ import it.unive.lisa.program.cfg.statement.call.Call;
 import it.unive.lisa.program.cfg.statement.call.UnresolvedCall;
 import it.unive.lisa.symbolic.SymbolicExpression;
 import it.unive.lisa.symbolic.value.ValueExpression;
-import java.util.Collection;
-import org.junit.Test;
 
 public class InformationFlowTest extends AnalysisTestExecutor {
 
@@ -138,7 +140,7 @@ public class InformationFlowTest extends AnalysisTestExecutor {
 		conf.serializeResults = true;
 		conf.abstractState = new SimpleAbstractState<>(
 				new MonolithicHeap(),
-				new InferenceSystem<>(new NonInterference()),
+				new NonInterference(),
 				new TypeEnvironment<>(new InferredTypes()));
 		conf.semanticChecks.add(new NICheck());
 		conf.testDir = "non-interference/confidentiality";
@@ -154,7 +156,7 @@ public class InformationFlowTest extends AnalysisTestExecutor {
 		conf.serializeResults = true;
 		conf.abstractState = new SimpleAbstractState<>(
 				new MonolithicHeap(),
-				new InferenceSystem<>(new NonInterference()),
+				new NonInterference(),
 				new TypeEnvironment<>(new InferredTypes()));
 		conf.semanticChecks.add(new NICheck());
 		conf.testDir = "non-interference/integrity";
@@ -170,7 +172,7 @@ public class InformationFlowTest extends AnalysisTestExecutor {
 		conf.serializeResults = true;
 		conf.abstractState = new SimpleAbstractState<>(
 				new MonolithicHeap(),
-				new InferenceSystem<>(new NonInterference()),
+				new NonInterference(),
 				new TypeEnvironment<>(new InferredTypes()));
 		conf.callGraph = new RTACallGraph();
 		conf.interproceduralAnalysis = new ContextBasedAnalysis<>(FullStackToken.getSingleton());
@@ -185,15 +187,12 @@ public class InformationFlowTest extends AnalysisTestExecutor {
 	private static class NICheck
 			implements
 			SemanticCheck<
-					SimpleAbstractState<MonolithicHeap, InferenceSystem<NonInterference>,
-							TypeEnvironment<InferredTypes>>> {
+					SimpleAbstractState<MonolithicHeap, NonInterference, TypeEnvironment<InferredTypes>>> {
 
 		@Override
-		@SuppressWarnings({ "unchecked" })
 		public boolean visit(
 				CheckToolWithAnalysisResults<
-						SimpleAbstractState<MonolithicHeap, InferenceSystem<NonInterference>,
-								TypeEnvironment<InferredTypes>>> tool,
+						SimpleAbstractState<MonolithicHeap, NonInterference, TypeEnvironment<InferredTypes>>> tool,
 				CFG graph,
 				Statement node) {
 			if (!(node instanceof Assignment))
@@ -206,18 +205,18 @@ public class InformationFlowTest extends AnalysisTestExecutor {
 				try {
 					AnalyzedCFG<?> result = (AnalyzedCFG<?>) res;
 					AnalysisState<?> post = result.getAnalysisStateAfter(assign);
-					InferenceSystem<NonInterference> state = post.getState().getDomainInstance(InferenceSystem.class);
+					NonInterference state = post.getState().getDomainInstance(NonInterference.class);
 					AnalysisState<?> postL = result.getAnalysisStateAfter(assign.getLeft());
-					InferenceSystem<NonInterference> left = postL.getState().getDomainInstance(InferenceSystem.class);
+					NonInterference left = postL.getState().getDomainInstance(NonInterference.class);
 					AnalysisState<?> postR = result.getAnalysisStateAfter(assign.getRight());
-					InferenceSystem<NonInterference> right = postR.getState().getDomainInstance(InferenceSystem.class);
+					NonInterference right = postR.getState().getDomainInstance(NonInterference.class);
 
 					for (SymbolicExpression l : postL.getState().rewrite(postL.getComputedExpressions(), assign,
 							postL.getState()))
 						for (SymbolicExpression r : postR.getState().rewrite(postR.getComputedExpressions(), assign,
 								postR.getState())) {
-							NonInterference ll = left.eval((ValueExpression) l, assign.getLeft(), postL.getState());
-							NonInterference rr = right.eval((ValueExpression) r, assign.getRight(), postR.getState());
+							NI ll = left.eval((ValueExpression) l, assign.getLeft(), postL.getState());
+							NI rr = right.eval((ValueExpression) r, assign.getRight(), postR.getState());
 
 							if (ll.isLowConfidentiality() && rr.isHighConfidentiality())
 								tool.warnOn(assign,
